# Refactor: trade slightly higher execution time for less memory allocation