@yunginnanet 2024-06-26
superseded by #14 which is free lunch comparatively